+++ test/lib/solid_session_test.rb
@@ -1,4 +1,0 @@
-require "test_helper"
-
-class SolidSessionTest < ActiveSupport::TestCase
-end