+++ test/lib/stored_session/log_subscriber_test.rb
@@ -1,0 +1,4 @@
+require "test_helper"
+
+class StoredSession::LogSubscriberTest < ActiveSupport::TestCase
+end