+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-# Solid Session Changelog
+# Stored Session Changelog
 
 ## Unreleased
 
-## [0.1.0](https://github.com/tbhb/solid_session/releases/tag/v0.1.0)
+## [0.2.0](https://github.com/tbhb/stored_session/releases/tag/v0.2.0)
+
+- Rename gem to Stored Session
+
+## [0.1.0](https://github.com/tbhb/stored_session/releases/tag/v0.1.0)
 
 - Initial release